[WIP]ajax・フラッシュメッセージの実装

+++ db/migrate/20220915074102_add_column_average_score_to_teacher.rb
@@ -1,5 +1,5 @@
 class AddColumnAverageScoreToTeacher < ActiveRecord::Migration[6.0]
   def change
-    add_column :teachers, :average_score, :integer
+    add_column :teachers, :average_score, :float
   end
 end
